完成退出显示 Signed-off-by: Frosty <3259024151@qq.com>

+++ src/main/java/UserFiles/TimeList.java
@@ -25,7 +25,7 @@ public class TimeList implements Serializable {
         playSec = 0;
     }
 
-    public void addSec(int sec) {
+    public void addSec(long sec) {
         playSec += sec;
     }
